% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{coef}
\alias{coef}
\title{coef}
\usage{
coef(object, model = c("full", "cutpoints", "inflation", "ordered"), ...)
}
\arguments{
\item{object}{an object of class \code{iop.mod} (output of \code{\link{iop}}).}
}
\description{
A function to extract coefficients from \code{iop} model results.
}
\examples{
 model1 <- iop(Y ~ X1 + X2|Z1 + Z2, data=data, type=c('ziop'))
 coef(model1)

}
