% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{fitted}
\alias{fitted}
\title{fitted}
\usage{
fitted(
  object,
  newdata = NULL,
  type = c("response.full", "response.ordered", "response.inflation", "linear"),
  ...
)
}
\arguments{
\item{object}{an object of class \code{iop.mod} (output of \code{\link{iop}}).}

\item{newdata}{An optional data frame in which to look for variables to use when model fitting.}

\item{type}{the tye of equation to be fitted. Options include "response.full" (both inflation and ordered probit stages), "response.ordered" (ordered probit stage only) and "response.inflation" (inflation stage only.}
}
\description{
A function that extracts fitted values from an object of class \code{iop.mod}.
}
\examples{
 model1 <- iop(Y ~ X1 + X2|Z1 + Z2, data=data, type=c('ziop'))
 fitted(model1, type=c("response.full"))

}
