% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{iop.mod.default}
\alias{iop.mod.default}
\title{iop.mod.default}
\usage{
iop.mod.default(x, z, y, weights, offsetx, offsetz, na.action, type, ...)
}
\arguments{
\item{object}{an object of class \code{iop.mod} (output of \code{\link{iop}}).}
}
\description{
Default method for a \code{ \link{iop}}.
}
