% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{iop.mod.Est}
\alias{iop.mod.Est}
\title{iop.mod.Est}
\usage{
iop.mod.Est(x, z, y, weights, offsetx, offsetz, na.action, type)
}
\arguments{
\item{x}{covariates for the ordered probit stage.}

\item{z}{covariates for the inflation (split population) stage.}

\item{y}{the ordinal dependent variable.}

\item{weights}{an optional vector of weights to be used in the fitting process.}

\item{offsetx}{offest value for the ordered probit stage covariates (X). This can be used to specify an a priori known component to be included in the linear predictor during fitting. For more information, see \code{\link{offset}}.}

\item{offsetz}{offest value for the inflation (split population) stage covariates (Z). This can be used to specify an a priori known component to be included in the linear predictor during fitting. For more information, see \code{\link{offset}}.}

\item{na.action}{a function indicating what should happen when NAs are included in the data. Options are "na.omit" or "na.fail". The default is "na.omit".}

\item{type}{type of inflation ordered probit model to be used. Options are "ziop" or "miop". The type of the inflation model must be specified.}
}
\description{
Raw form of the \code{ \link{iop}} function. For user-friendly formula-oriented command, use \code{ \link{iop}}.
}
