% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{iop}
\alias{iop}
\title{iop}
\usage{
iop(
  formula,
  data = list(),
  weights = NULL,
  offset = NULL,
  na.action = c("na.omit", "na.fail"),
  type = c("ziop", "miop"),
  ...
)
}
\arguments{
\item{formula}{a formula in the form Y ~ X1 + X2... | Z1 + Z2 ... where Y is the ordered probit dependent variable; Xi are the ordered probit stage covariates; and the Zi are the inflation (split population) stage covariates. See \code{link{formula}}.}

\item{data}{list object of data.}

\item{weights}{an optional vector of weights to be used in the fitting process. Default is NULL.}

\item{offset}{This can be used to specify an a priori known component to be included in the linear predictor during fitting. The same offset is applied to both stages. See \code{\link{offset}}.}

\item{na.action}{a function indicating what should happen when NAs are included in the data. Options are "na.omit" or "na.fail". The default is "na.omit".}

\item{type}{type of inflation ordered probit model to be used. Options are "ziop" or "miop". The type of the inflation model must be specified.}
}
\description{
\code{iop} fits an ordered probit model with inflations in either the "zero (bottom)" or "middle" categories.
}
\examples{
 model1 <- iop(Y ~ X1 + X2|Z1 + Z2, data=data, type=c('ziop'))

}
