% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{miop}
\alias{miop}
\title{miop}
\usage{
miop(est, Y, X, Z, data, weights, offsetx, offsetz)
}
\arguments{
\item{est}{starting values for the estimation. Vector of length of the number of parameters.}

\item{Y}{the ordinal dependent variable.}

\item{X}{covariates for the ordered probit stage.}

\item{Z}{covariates for the inflation (split population) stage.}

\item{data}{dataset that contains the dependent and independent variables.}

\item{weights}{an optional vector of weights to be used in the fitting process.}

\item{offsetx}{offest value for the ordered probit stage covariates (X). This can be used to specify an a priori known component to be included in the linear predictor during fitting. For more information, see \code{\link{offset}}.}

\item{offsetz}{offest value for the inflation (split population) stage covariates (Z). This can be used to specify an a priori known component to be included in the linear predictor during fitting. For more information, see \code{\link{offset}}.}
}
\value{
Likelihood of the MIOP model specification.
}
\description{
Likelihood function for MIOP model.
}
