% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{predict}
\alias{predict}
\title{predict}
\usage{
predict(
  object,
  newdata = NULL,
  type = c("prob.full", "prob.ordered", "prob.inflation", "response.full",
    "response.ordered", "response.inflation", "linear"),
  ...
)
}
\arguments{
\item{object}{an object of class \code{iop.mod} (output of \code{\link{iop}}).}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{type}{}
}
\description{
A function to extract predicted values from the IOP model object.
}
\examples{
 model1 <- iop(Y ~ X1 + X2|Z1 + Z2, data=data, type=c('ziop'))
 predict(model1, type=c("response.full"))

}
