% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{print.summary.iop.mod}
\alias{print.summary.iop.mod}
\title{print.iop.mod}
\usage{
\method{print}{summary.iop.mod}(object, ...)
}
\arguments{
\item{object}{an object of class \code{iop.mod} (output of \code{\link{iop}}).}
}
\description{
Print method for a \code{ \link{iop}} object.
}
