% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{residuals}
\alias{residuals}
\title{residuals}
\usage{
residuals(object, type = c("response"), ...)
}
\arguments{
\item{object}{an object of class \code{iop.mod} (output of \code{\link{iop}}).}
}
\description{
A function to extract residuals from \code{iop} model results.
}
\examples{
 model1 <- iop(Y ~ X1 + X2|Z1 + Z2, data=data, type=c('ziop'))
 residuals(model1)

}
