% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{summary.iop.mod}
\alias{summary.iop.mod}
\title{summary.iop.mod}
\usage{
\method{summary}{iop.mod}(object, ...)
}
\arguments{
\item{object}{an object of class \code{iop.mod} (output of \code{\link{iop}}.}
}
\description{
Summary method for a \code{ \link{iop}} object.
}
