% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmiop.R
\name{vcov}
\alias{vcov}
\title{vcov}
\usage{
vcov(object, model = c("full", "cutpoints", "inflation", "ordered"), ...)
}
\arguments{
\item{object}{an object of class \code{iop.mod} (output of \code{\link{iop}}).}
}
\description{
A function to extract variance-covariance matrix from \code{iop} model results.
}
\examples{
 model1 <- iop(Y ~ X1 + X2|Z1 + Z2, data=data, type=c('ziop'))
 vcov(model1)

}
